requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.5
django-cors-headers==4.2.0
django-jsonfield==1.4.1
django-telegram==0.1.0
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
Pillow==10.0.1
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
yarl==1.9.2
